Flex driver should handle environments without syslog
OCP controller manager containers haven't got syslog installed.
To handle that situation and unlock this driver to work in such
environment, we need to redirect logs to another destination.

We were experimenting with using stdout/stderr to achieve that
goal but, OCP is parsing whole output(i.e. incl. both stdout and
stderr). To don't loose logs we've decided to use logging to file.